requirements.txt
numpy>=1.19
pandas>=1.1
pyyaml==6.0.0
tf-models-official==2.14.0
ml_collections==0.1.1
absl-py>=0.13.0
protobuf>=3.20.3
flatbuffers>=23.1.21
keras<2.15,>=2.14.0
tensorflow-estimator<2.15,>=2.14.0
zipp>=3.1.0
httplib2>=0.15.0
httplib2<1dev,>=0.15.0
pysam
testresources
intervaltree
